renamed 'index' to 'fetcher'
