+++ apps/dataset-browser/src/lib/dataset-fetcher/fetcher-result.ts
